Error Handling

+++ model/StoreInfo.js
@@ -141,6 +141,24 @@ const storeInfo = {
 
     delete: (storeIdx) => {
         return new Promise(async (resolve, reject) => {
+            const storeIdxQuery = `SELECT * FROM store_info WHERE store_idx = ?`;
+            const storeIdxResult = await pool.queryParam_Arr(storeIdxQuery, [storeIdx]);
+
+            if (!storeIdxResult) {
+                resolve({
+                    code : statusCode.INTERNAL_SERVER_ERROR,
+                    json : authUtil.successFalse(statusCode.INTERNAL_SERVER_ERROR, responseMessage.INTERNAL_SERVER_ERROR)
+                });
+                return;
+            }
+            if (storeIdxResult[0] == undefined) {
+                resolve({
+                    code : statusCode.BAD_REQUEST,
+                    json : authUtil.successFalse(statusCode.BAD_REQUEST, `존재하지 않는 상점입니다`)
+                });
+                return;
+            }
+            
             const deleteStoreQuery = `DELETE FROM ${storeInfoTable} WHERE store_idx = ?`;
             const deleteStoreResult = await pool.queryParam_Arr(deleteStoreQuery, [storeIdx]);
